카테고리뷰 삭제버튼 수정
